use size_t

+++ src/acl/external/time_quota/ext_time_quota_acl.cc
@@ -56,7 +56,7 @@ const char* KeyTimeBudgetLeft = "time-budget-left";
 const char* KeyTimeBudgetConfigured = "time-budget-configured";
 
 /** Maximum size of buffers used to read or display lines. */
-const int TQ_BUFFERSIZE = 1024;
+const size_t TQ_BUFFERSIZE = 1024;
 
 /** If there is more than this given number of seconds between two
  * successive requests, than the second request will be treated as a
